Refactoring OfferDetails screen to make tests
