+++ mobile/screens/tab_navigator/vendas/screens/offers/OfferDetails.js
@@ -12,73 +12,7 @@ import { Card, CardItem, Text, Left, Right, Content, Body} from 'native-base';
 import OfferDialog from '../../components/OfferDialog';
 import jwt_decode from 'jwt-decode';
 import GreenButton from '../../components/GreenButton';
-
-const Quantity = [
-  {
-    label: '1',
-    value: '1'
-  },
-  {
-    label: '2',
-    value: '2'
-  },
-  {
-    label: '3',
-    value: '3'
-  },
-  {
-    label: '4',
-    value: '4'
-  },
-  {
-    label: '5',
-    value: '5'
-  },
-  {
-    label: '6',
-    value: '6'
-  },
-  {
-    label: '7',
-    value: '7'
-  },
-  {
-    label: '8',
-    value: '8'
-  },
-  {
-    label: '9',
-    value: '9'
-  },
-  {
-    label: '10',
-    value: '10'
-  }
-];
-
-class FormPicker extends Component {
-  constructor(props){
-    super(props);
-    this.state = {
-        modalVisible: false
-    };
-  }
-
-  render(){
-    return (
-      <Picker
-      selectedValue = {this.props.value}
-      onValueChange = {this.props.onValueChange}
-      style = {styles.PickerStyle}
-      mode = 'dropdown'
-      >
-      {this.props.items.map((i, index) => (
-        <Picker.Item key = {index} label = {i.label} value = {i.value} />
-      ))}
-      </Picker>
-    );
-  }
-}
+import FormPicker from '../../components/FormPicker';
 
 
  class OfferDetails extends Component {
@@ -192,7 +126,6 @@ class FormPicker extends Component {
               <Text style = {styles.PickerText}> Quantidade </Text>
 
               <FormPicker
-                items={Quantity}
                 value={this.state.quantity}
                 onValueChange={(itemValue, itemIndex) => this.setState({ quantity: itemValue })}
               />
